such a rename
Is this name better than `_angular_momentum_generators`?
