+++ src/openfermion/hamiltonians/__init__.py
@@ -10,13 +10,13 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from ._spin_operators import (up_index, down_index,
-                              s_squared_operator, s_plus_operator,
-                              s_minus_operator, sz_operator)
-
 from ._chemical_series import (make_atomic_ring,
                                make_atomic_lattice,
                                make_atom)
+
+from ._fermion_spin_operators import (up_index, down_index,
+                                      s_squared_operator, s_plus_operator,
+                                      s_minus_operator, sz_operator)
 
 from ._hubbard import fermi_hubbard
