+++ src/openfermion/hamiltonians/_fermion_spin_operators.py
